+++ odrivemotor/src/odriveSerial/__init__.py
@@ -1,11 +1,9 @@
 """
-This files odrive model with the Viam Registry.
+This files OdriveSerial model with the Viam Registry.
 """
 
 from viam.resource.registry import Registry, ResourceCreatorRegistration
 from viam.components.motor import Motor
 from .odriveSerial.odriveSerial import OdriveSerial
-from .odriveCAN.odriveCAN import OdriveCAN
 
 Registry.register_resource_creator(Motor.SUBTYPE, OdriveSerial.MODEL, ResourceCreatorRegistration(OdriveSerial.new, OdriveSerial.validate))
-Registry.register_resource_creator(Motor.SUBTYPE, OdriveCAN.MODEL, ResourceCreatorRegistration(OdriveCAN.new, OdriveCAN.validate))
